+++ src/main/java/com/senecafoundation/Controllers/ChoiceController.java
@@ -1,9 +1,12 @@
-package com.senecafoundation;
-import java.util.List;
-import org.springframework.web.bind.annotation.PathVariable;
+package com.senecafoundation.Controllers;
 import java.util.UUID;
+import java.util.List;
+import com.senecafoundation.DataHandler.ChoiceDataHandler;
 import com.senecafoundation.DataHandler.ResponseDataHandler;
+import com.senecafoundation.Scene.Choice;
 import com.senecafoundation.Scene.Response;
+
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,58 +18,65 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping("response")
-public class ResponseController 
+@RequestMapping("choice")
+public class ChoiceController 
 {
     @Autowired
-    ResponseDataHandler<Response> dataHandler;
+    ChoiceDataHandler<Choice> dataHandler;
+    @Autowired
+    ResponseDataHandler<Response> responseDataHandler;
 
     @GetMapping("/createform")
     public String showForm(Model model) {
-        Response response = new Response();
-        model.addAttribute("response", response);
-        return "create_response";
+        Choice choice = new Choice();
+        List<Response> responses = responseDataHandler.ReadAll();
+        model.addAttribute("choice", choice);
+        model.addAttribute("possibleResponses", responses);
+        return "create_choice";
     }
+    
 
     @RequestMapping(value = "/createform", method = RequestMethod.POST)
-    public String submit(@ModelAttribute("response") Response response, BindingResult result, ModelMap model) {
+    public String submit(@ModelAttribute("choice") Choice choice, BindingResult result, ModelMap model) {
         if (result.hasErrors()) {
             return "error";
         }
-        dataHandler.Create(response);
+        dataHandler.Create(choice);
         //repo.save(shadowElf);
-        model.addAttribute("response", response);
-        return "response";
+        model.addAttribute("choice", choice);
+        return "choice";
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public String showFormRead(@PathVariable("id") String Id, ModelMap model) throws Exception {
-        Response response = (Response) dataHandler.Read(UUID.fromString(Id));
-        model.addAttribute("response",response);
-        return "response";
+        Choice choice = (Choice) dataHandler.Read(UUID.fromString(Id));
+        model.addAttribute("choice",choice);
+        return "choice";
     }
 
     @RequestMapping(value ="/updateform/{id}", method = RequestMethod.GET)
     public String showUpdateForm(@PathVariable("id") String Id, Model model) throws Exception {
-        Response response = (Response) dataHandler.Read(UUID.fromString(Id));
-        model.addAttribute("response", response);
-        return "create_response"; 
+        Choice choice = (Choice) dataHandler.Read(UUID.fromString(Id));
+        List<Response> responses = responseDataHandler.ReadAll();
+        model.addAttribute("choice", choice);
+        model.addAttribute("possibleResponses", responses);
+        return "create_choice"; 
     }
     @RequestMapping(value="/updateform", method = RequestMethod.POST)
-    public String change(Response response, BindingResult result, ModelMap model) {
+    public String change(Choice choice, BindingResult result, ModelMap model) {
         if (result.hasErrors()) {
             return "error";
         }
-        dataHandler.Update(response);
-        model.addAttribute("response",response);
-        return "response";   
+        dataHandler.Update(choice);
+        model.addAttribute("choice",choice);
+        return "choice";   
     }
 
     @GetMapping("/deleteform")
     public String showFormDelete(Model model) {
-        List<Response> responseList = dataHandler.ReadAll();
-        model.addAttribute("response", responseList);
-        return "delete_Response";//connects to html file
+        List<Choice> choiceList = dataHandler.ReadAll();
+        model.addAttribute("choice", choiceList);
+        return "delete_Choice";//connects to html file
     }
 
     @RequestMapping(value = "/deleteform/{id}", method = RequestMethod.DELETE)
