+++ src/main/java/com/senecafoundation/Controllers/ResponseController.java
@@ -1,12 +1,9 @@
-package com.senecafoundation;
-import java.util.UUID;
+package com.senecafoundation.Controllers;
 import java.util.List;
-import com.senecafoundation.DataHandler.ChoiceDataHandler;
-import com.senecafoundation.DataHandler.ResponseDataHandler;
-import com.senecafoundation.Scene.Choice;
-import com.senecafoundation.Scene.Response;
-
 import org.springframework.web.bind.annotation.PathVariable;
+import java.util.UUID;
+import com.senecafoundation.DataHandler.ResponseDataHandler;
+import com.senecafoundation.Scene.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,65 +15,58 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping("choice")
-public class ChoiceController 
+@RequestMapping("response")
+public class ResponseController 
 {
     @Autowired
-    ChoiceDataHandler<Choice> dataHandler;
-    @Autowired
-    ResponseDataHandler<Response> responseDataHandler;
+    ResponseDataHandler<Response> dataHandler;
 
     @GetMapping("/createform")
     public String showForm(Model model) {
-        Choice choice = new Choice();
-        List<Response> responses = responseDataHandler.ReadAll();
-        model.addAttribute("choice", choice);
-        model.addAttribute("possibleResponses", responses);
-        return "create_choice";
+        Response response = new Response();
+        model.addAttribute("response", response);
+        return "create_response";
     }
-    
 
     @RequestMapping(value = "/createform", method = RequestMethod.POST)
-    public String submit(@ModelAttribute("choice") Choice choice, BindingResult result, ModelMap model) {
+    public String submit(@ModelAttribute("response") Response response, BindingResult result, ModelMap model) {
         if (result.hasErrors()) {
             return "error";
         }
-        dataHandler.Create(choice);
+        dataHandler.Create(response);
         //repo.save(shadowElf);
-        model.addAttribute("choice", choice);
-        return "choice";
+        model.addAttribute("response", response);
+        return "response";
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public String showFormRead(@PathVariable("id") String Id, ModelMap model) throws Exception {
-        Choice choice = (Choice) dataHandler.Read(UUID.fromString(Id));
-        model.addAttribute("choice",choice);
-        return "choice";
+        Response response = (Response) dataHandler.Read(UUID.fromString(Id));
+        model.addAttribute("response",response);
+        return "response";
     }
 
     @RequestMapping(value ="/updateform/{id}", method = RequestMethod.GET)
     public String showUpdateForm(@PathVariable("id") String Id, Model model) throws Exception {
-        Choice choice = (Choice) dataHandler.Read(UUID.fromString(Id));
-        List<Response> responses = responseDataHandler.ReadAll();
-        model.addAttribute("choice", choice);
-        model.addAttribute("possibleResponses", responses);
-        return "create_choice"; 
+        Response response = (Response) dataHandler.Read(UUID.fromString(Id));
+        model.addAttribute("response", response);
+        return "create_response"; 
     }
     @RequestMapping(value="/updateform", method = RequestMethod.POST)
-    public String change(Choice choice, BindingResult result, ModelMap model) {
+    public String change(Response response, BindingResult result, ModelMap model) {
         if (result.hasErrors()) {
             return "error";
         }
-        dataHandler.Update(choice);
-        model.addAttribute("choice",choice);
-        return "choice";   
+        dataHandler.Update(response);
+        model.addAttribute("response",response);
+        return "response";   
     }
 
     @GetMapping("/deleteform")
     public String showFormDelete(Model model) {
-        List<Choice> choiceList = dataHandler.ReadAll();
-        model.addAttribute("choice", choiceList);
-        return "delete_Choice";//connects to html file
+        List<Response> responseList = dataHandler.ReadAll();
+        model.addAttribute("response", responseList);
+        return "delete_Response";//connects to html file
     }
 
     @RequestMapping(value = "/deleteform/{id}", method = RequestMethod.DELETE)
